requirements-dev.txt
-r requirements.txt
ipython
jupyter
django-test-without-migrations
